Move AtFwd hal from device to framework manifest.
AtFwd hal is implemented as an android service and
it stars from system partition, that is, it's not
a traditional vendor hal and hence it should be
declared in the framework manifest and not in the
device manifest.
This commit introduces new device specific framework
manifest (framework_manifest.xml) and moves AtFwd hal
from device to framework manifest.

Change-Id: I6133df9199e419d1a1edb9535d2bf8b20e0d9bea
CRs-Fixed: 2287149